requirements.txt
streamlit
openai
python-dotenv
yt-dlp
trafilatura
requests
beautifulsoup4
pydub
langchain
langchain-openai
langchain-community
tiktoken
unstructured
nltk
faiss-cpu
langchain-chroma